fix stage c weight is loaded in bf16/fp16 #1119

+++ stable_cascade_train_stage_c.py
@@ -137,7 +137,7 @@ def train(args):
     # モデルを読み込む
     loading_device = accelerator.device if args.lowram else "cpu"
     effnet = sc_utils.load_effnet(args.effnet_checkpoint_path, loading_device)
-    stage_c = sc_utils.load_stage_c_model(args.stage_c_checkpoint_path, dtype=weight_dtype, device=loading_device)
+    stage_c = sc_utils.load_stage_c_model(args.stage_c_checkpoint_path, device=loading_device)  # dtype is as it is
     text_encoder1 = sc_utils.load_clip_text_model(args.text_model_checkpoint_path, dtype=weight_dtype, device=loading_device)
 
     if args.sample_at_first or args.sample_every_n_steps is not None or args.sample_every_n_epochs is not None:
